// src/features/Todo/interfaces/cons.ts
export enum EActionType {
  AllTodo = "allTodo",
  AddTodo = "addTodo",
  DeleteTodo = "deleteTodo",
  EditTodo = "editTodo",
  Loading = "loading",
  // TickTodo = "tickTodo",
  Fail = "fail",
  Succsess = "succsess",
}

export enum ETodoStatus {
  Active = "Active",
  Completed = "Completed",
  All = "All",
  FilterName = "filter-name",
  IsActive = "isActive",
}